+++ lib/libesp32/ESP32-to-ESP8266-compat/src/ESP8266WiFi.h
@@ -37,6 +37,11 @@ typedef enum WiFiPhyMode
 class WiFiClass32 : public WiFiClass
 {
 public:
+    wl_status_t begin(const char* wpa2_ssid, wpa2_auth_method_t method, const char* wpa2_identity=NULL, const char* wpa2_username=NULL, const char *wpa2_password=NULL, const char* ca_pem=NULL, const char* client_crt=NULL, const char* client_key=NULL, int32_t channel=0, const uint8_t* bssid=0, bool connect=true);
+    wl_status_t begin(const char* ssid, const char *passphrase = NULL, int32_t channel = 0, const uint8_t* bssid = NULL, bool connect = true);
+    wl_status_t begin(char* ssid, char *passphrase = NULL, int32_t channel = 0, const uint8_t* bssid = NULL, bool connect = true);
+    wl_status_t begin();
+
     static void hostname(const char* aHostname)
     {
         WiFi.setHostname(aHostname);
@@ -51,7 +56,14 @@ public:
     static void forceSleepWake();
     static bool getNetworkInfo(uint8_t i, String &ssid, uint8_t &encType, int32_t &rssi, uint8_t* &bssid, int32_t &channel, bool &hidden_scan);
 
-    bool IPv6(bool state);          // make sure it always exists even with older Arduino framework
+    static void dnsDone(void);      // used by the callback to stop the dns timer
+    int hostByName(const char* aHostname, IPAddress& aResult, int32_t timer_ms);
+    int hostByName(const char* aHostname, IPAddress& aResult);
+
+    void saveDNS(void);
+    void restoreDNS(void);
+protected:
+    ip_addr_t dns_save[DNS_MAX_SERVERS] = {};
 };
 
 void wifi_station_disconnect();